Item SO
Set up an item SO system for consumables that would be in the item shop and maybe inventory. You simply put the prefab into the shop window, and then you can set it to whatever scriptable object (that is a consumable) you like and it will display the information. Actual money, purchasing system, ect is not done yet.